requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
mongoengine==0.29.3
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.2
python-ulid==3.1.0
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
